Updated layout, renamed database to Software4You, added flag for resetting database

+++ Startup.cs
@@ -60,10 +60,15 @@ namespace Layout
                     pattern: "{controller=Home}/{action=Index}/{id?}");
             });
 
-            db.Database.EnsureDeleted();    // wipe out existing database
-            db.Database.EnsureCreated();    // our database is created after this line
 
-            new DbSeedData(db).Seed();
+            bool wantreset = Configuration.GetValue<bool>("Db:WantReset");
+            if(wantreset)
+            {
+                db.Database.EnsureDeleted();    // wipe out existing database
+                db.Database.EnsureCreated();    // our database is created after this line
+
+                new DbSeedData(db).Seed();
+            }
         }
     }
 }
